Update README.md
Set up a Node.js project: To begin, you create a new folder on your computer where you'll store your Node.js project. This folder will serve as the root directory for your project. You navigate to this folder using the command line or terminal.

Install the necessary dependencies: In order to interact with the file system in Node.js, you need to use the built-in fs module. This module provides methods for reading, writing, and manipulating files and directories. To use this module, you install it as a dependency for your project using the Node Package Manager (NPM). Running the npm install fs command fetches and installs the fs module into your project.

Create a JavaScript file: Inside your project folder, you create a new file and give it a meaningful name, such as deleteFiles.js. This file will contain the code that deletes the files in the specified folder.

Import the 'fs' module: To utilize the functionality provided by the fs module, you import it into your JavaScript file using the require() function. This allows you to access the methods and properties of the fs module within your script.

Define the folder path: You need to specify the path to the folder that contains the files you want to delete. This can be done by assigning the folder path to a variable in your script. The path can be relative or absolute, depending on the location of the folder in relation to your script.

Delete the files: In this step, you leverage the methods provided by the fs module to delete the files within the specified folder. Firstly, you use the readdirSync() function to read the contents of the folder and retrieve an array of file names. Then, you iterate over each file name using the forEach() method and construct the full file path by combining the folder path and the file name. Finally, you use the unlinkSync() function to delete each file using its respective file path.

Run the script: After saving your changes to the deleteFiles.js script, you execute it by running the command node deleteFiles.js in the command line or terminal. This command instructs Node.js to execute the script and carry out the file deletion process.